# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FichoChat API",
        "description": "API REST versionnée de FichoChat. Chaque endpoint documente des exemples JSON complets de réponses (succès et erreurs).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/conversations/{conversationId}/messages": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Lister les messages d’une conversation",
                "description": "Retourne ciphertext + iv (jamais de clair). Pagination cursor via `before` (UUID) + `limit`.",
                "operationId": "messagesIndex",
                "parameters": [
                    {
                        "name": "conversationId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "before",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 50,
                            "maximum": 100
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessagesIndex200"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthorized401"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Email non vérifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorForbidden403"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Conversation introuvable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorNotFound404"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorServer500"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "Envoyer un message chiffré",
                "description": "Le body contient ciphertext + iv + enveloppes CEK. Aucun texte en clair.",
                "operationId": "messagesStore",
                "parameters": [
                    {
                        "name": "conversationId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "sender_device_id",
                                    "ciphertext",
                                    "iv",
                                    "envelopes"
                                ]
                            },
                            "example": {
                                "sender_device_id": "019fc7e1-aaaa-bbbb-cccc-ddddeeeeffff",
                                "ciphertext": "base64…",
                                "iv": "base64…",
                                "type": "text",
                                "envelopes": [
                                    {
                                        "recipient_device_id": "019fc7e1-1111-2222-3333-444455556666",
                                        "encrypted_cek": "base64…",
                                        "ephemeral_public_key": "-----BEGIN PUBLIC KEY-----\\n…",
                                        "iv": "base64…"
                                    }
                                ],
                                "media_ids": []
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessagesStore201"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthorized401"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Email non vérifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorForbidden403"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Conversation introuvable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorNotFound404"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorValidation422"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorServer500"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/messages/{id}/for-me": {
            "delete": {
                "tags": [
                    "Messages"
                ],
                "summary": "Supprimer un message pour soi uniquement",
                "description": "Le message disparaît de la liste pour l’utilisateur courant ; l’autre participant le conserve.",
                "operationId": "messagesDestroyForMe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessagesDeleteForMe200"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthorized401"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Email non vérifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorForbidden403"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Introuvable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorNotFound404"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorValidation422"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorServer500"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/messages/{id}/receipts": {
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "Accusé delivered / read",
                "operationId": "messagesReceipt",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ]
                            },
                            "example": {
                                "status": "read",
                                "device_id": "019fc7e1-aaaa-bbbb-cccc-ddddeeeeffff"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessagesReceipt200"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthorized401"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Email non vérifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorForbidden403"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Introuvable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorNotFound404"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorValidation422"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorServer500"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorUnauthorized401": {
                "description": "Non authentifié",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object",
                "example": {
                    "success": false,
                    "message": "Unauthenticated."
                }
            },
...
            "MessagesIndex200": {
                "type": "object",
                "example": {
                    "success": true,
                    "message": "Operation completed successfully.",
                    "data": {
                        "messages": [
                            {
                                "id": "019fcb01-aaaa-bbbb-cccc-ddddeeeeffff",
                                "conversation_id": "019fca01-aaaa-bbbb-cccc-ddddeeeeffff",
                                "sender_user_id": "019fc7e1-c324-7380-a855-84056ddd751c",
                                "sender_device_id": "019fc7e1-aaaa-bbbb-cccc-ddddeeeeffff",
                                "type": "text",
                                "ciphertext": "base64…",
                                "iv": "base64…",
                                "reply_to_message_id": null,
                                "edited_at": null,
                                "deleted_for_everyone": false,
                                "receipts": [],
                                "medias": [],
                                "created_at": "2026-08-03T20:20:00.000000Z",
                                "updated_at": "2026-08-03T20:20:00.000000Z"
                            }
                        ]
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Access Token obtenu via /auth/login ou /auth/register. Header: Authorization: Bearer {access_token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}